python-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
chromadb==0.4.18
pypdf==3.17.4
python-docx==1.1.0
python-multipart==0.0.6
pydantic==2.5.2
python-dotenv==1.0.0
supabase==2.3.0
sentence-transformers==2.2.2
tiktoken==0.5.2